OTAT-464: Fix tests and clean up services

+++ src/main/java/org/flickit/flickitassessmentcore/domain/Assessment.java
@@ -19,7 +19,7 @@ public class Assessment {
     private LocalDateTime creationTime;
     private LocalDateTime lastModificationDate;
     private Long assessmentKitId;
-    private Long colorId;
+    private Integer colorId;
     private Long spaceId;
     private MaturityLevel maturityLevel;
